+++ app/Propstack/Fields.php
@@ -125,6 +125,7 @@ class Fields {
 		$field->set_columns(
 			array(
 				__( 'The field', 'connector-for-propstack' ),
+				__( 'Internal name', 'connector-for-propstack' ),
 				__( 'Example from your objects', 'connector-for-propstack' ),
 			)
 		);
@@ -155,6 +156,16 @@ class Fields {
 			$field->add_setting( $name_setting, $row, 0 );
 
 			// add setting.
+			$name_setting = $settings_obj->add_setting( 'propstack_connector_fields_all_' . $immo_field->get_name() . '_internal_name' );
+			$name_setting->prevent_export( true );
+			$name_setting->set_type( 'string' );
+			$name_setting->set_section( $hidden_section );
+			$name_setting_field = new Value( $settings_obj );
+			$name_setting_field->set_value( '<code data-copied-label="' . esc_attr__( 'copied', 'connector-for-propstack' ) . '" title="' . esc_attr__( 'Click to copy this code in your clipboard', 'connector-for-propstack' ) . '">' . $immo_field->get_name() . '</code>' );
+			$name_setting->set_field( $name_setting_field );
+			$field->add_setting( $name_setting, $row, 1 );
+
+			// add setting.
 			$name_setting = $settings_obj->add_setting( 'propstack_connector_fields_all_' . $immo_field->get_name() . '_example' );
 			$name_setting->prevent_export( true );
 			$name_setting->set_type( 'string' );
@@ -162,7 +173,7 @@ class Fields {
 			$name_setting_field = new Value( $settings_obj );
 			$name_setting_field->set_value( $fields_tab_called ? Cache::get( $immo_field->get_name() . '_example' ) : '' );
 			$name_setting->set_field( $name_setting_field );
-			$field->add_setting( $name_setting, $row, 1 );
+			$field->add_setting( $name_setting, $row, 2 );
 
 			// next row.
 			++$row;
@@ -285,6 +296,7 @@ class Fields {
 			$field->set_columns(
 				array(
 					__( 'The field', 'connector-for-propstack' ),
+					__( 'Internal name', 'connector-for-propstack' ),
 					__( 'Hide in frontend', 'connector-for-propstack' ),
 				)
 			);
@@ -314,6 +326,16 @@ class Fields {
 				$field->add_setting( $name_setting, $row, 0 );
 
 				// add setting.
+				$name_setting = $settings_obj->add_setting( 'propstack_connector_fields_' . $taxonomy->get_name() . '_' . $immo_field->get_name() . '_internal_name' );
+				$name_setting->prevent_export( true );
+				$name_setting->set_type( 'string' );
+				$name_setting->set_section( $hidden_section );
+				$name_setting_field = new Value( $settings_obj );
+				$name_setting_field->set_value( '<code data-copied-label="' . esc_attr__( 'copied', 'connector-for-propstack' ) . '" title="' . esc_attr__( 'Click to copy this code in your clipboard', 'connector-for-propstack' ) . '">' . $immo_field->get_name() . '</code>' );
+				$name_setting->set_field( $name_setting_field );
+				$field->add_setting( $name_setting, $row, 1 );
+
+				// add setting.
 				$disable_field_setting = $settings_obj->add_setting( 'propstack_connector_fields_' . $taxonomy->get_name() . '_' . $immo_field->get_name() . '_disabled' );
 				$disable_field_setting->set_default( $taxonomy->is_field_default_disabled( $immo_field ) ? 1 : 0 );
 				$disable_field_setting->set_section( $hidden_section );
@@ -321,7 +343,7 @@ class Fields {
 				$disable_field_setting_field->set_title( $immo_field->get_label() );
 				$disable_field_setting_field->set_readonly( true );
 				$disable_field_setting->set_field( $disable_field_setting_field );
-				$field->add_setting( $disable_field_setting, $row, 1 );
+				$field->add_setting( $disable_field_setting, $row, 2 );
 
 				// next row.
 				++$row;